+++ nip/parameters/agents.py
@@ -18,7 +18,11 @@ from nip.parameters.parameters_base import (
     register_parameter_class,
     register_parameter_value_class,
 )
-from nip.parameters.types import ActivationType, ImageBuildingBlockType
+from nip.parameters.types import (
+    ActivationType,
+    ImageBuildingBlockType,
+    UseSupervisorType,
+)
 from nip.parameters.update_schedule import (
     AgentUpdateSchedule,
     ConstantUpdateSchedule,
@@ -490,6 +494,17 @@ class PureTextAgentParameters(AgentParameters):
     system_prompt_template_path : str | None
         This option allows specifying a custom system prompt template. If not provided,
         the default system prompt template is used.
+    use_supervisor_message : UseSupervisorType
+        Whether and when to use a 'supervisor' message when generating responses. This
+        is a message which is appended to the chat history, with instructions for the
+        model. These instructions are already included in the system prompt, but this
+        can help improve the quality of the generated responses. Some models also
+        require at least one user message to be able to generate a response, and this
+        can be used to work around that. The options are listed in the
+        :const:`UseSupervisorType <nip.parameters.types.UseSupervisorType>` enum, and
+        specify when to use the supervisor message.
+    supervisor_name : str
+        The name of the user who sends the supervisor message.
     max_response_words : int
         In the system prompt, we say that the agent should respond with a message of at
         most this many words.
@@ -515,6 +530,9 @@ class PureTextAgentParameters(AgentParameters):
     freeze_agent: bool = False
 
     system_prompt_template_path: str | None = None
+
+    use_supervisor_message: UseSupervisorType = "none"
+    supervisor_name: str = "Supervisor"
 
     max_response_words: int = 150
 
@@ -543,10 +561,12 @@ class CodeValidationAgentParameters(PureTextAgentParameters):
 
     Parameters
     ----------
-    model_provider : Literal["OpenAI"]
+    model_provider : Literal["OpenAI", "vLLM-OpenAI", "OpenRouter"]
         The provider of the model and API to use.
     model_name : str
         The name of the model to use.
+    vllm_openai_base_url : str
+        When using vLLM's OpenAI-compatible server, this is the URL of the server
     use_dummy_api : bool
         Whether to use a dummy API instead of the real API. This is useful for testing
         the agent without making real API requests.
@@ -554,7 +574,8 @@ class CodeValidationAgentParameters(PureTextAgentParameters):
         The group of agents which share the same model. When two agents share this
         value, they will use the same model inference. For fine-tuning, this model is
         trained on a copy of the rollouts and rewards for each agent in the group. When
-        this is ``None``, the agent is in a group on its own.
+        this is ``None``, the agent is in a group whose name is the same as the agent's
+        name.
     temperature : float | None
         The temperature to use when sampling from the model. If ``None``, the model uses
         the default temperature. Only one of ``temperature`` and ``top_p`` should be
@@ -563,11 +584,30 @@ class CodeValidationAgentParameters(PureTextAgentParameters):
         The top-p value to use when sampling from the model. A value 0.1 means only the
         top 10% of tokens are considered when sampling. If ``None``, the model uses the
         default top-p value. Only one of ``temperature`` and ``top_p`` should be set.
+    repetition_penalty : float | None
+        Float that penalizes new tokens based on whether they appear in the prompt and
+        the generated text so far. Values > 1 encourage the model to use new tokens,
+        while values < 1 encourage the model to repeat tokens. Not all models support
+        this parameter.
     fine_tune_from_scratch : bool
         Whether to fine-tune the model from scratch each iteration, or continue
         fine-tuning from the previous iteration.
     freeze_agent : bool
         Whether to freeze the agent (i.e. not fine-tune it).
+    system_prompt_template_path : str | None
+        This option allows specifying a custom system prompt template. If not provided,
+        the default system prompt template is used.
+    use_supervisor_message : UseSupervisorType
+        Whether and when to use a 'supervisor' message when generating responses. This
+        is a message which is appended to the chat history, with instructions for the
+        model. These instructions are already included in the system prompt, but this
+        can help improve the quality of the generated responses. Some models also
+        require at least one user message to be able to generate a response, and this
+        can be used to work around that. The options are listed in the
+        :const:`UseSupervisorType <nip.parameters.types.UseSupervisorType>` enum, and
+        specify when to use the supervisor message.
+    supervisor_name : str
+        The name of the user who sends the supervisor message.
     max_response_words : int
         In the system prompt, we say that the agent should respond with a message of at
         most this many words.